+++ src/opensynth/models/faraday/__init__.py
@@ -1,7 +1,14 @@
 # Copyright Contributors to the Opensynth-energy Project.
 # SPDX-License-Identifier: Apache-2.0
 
+from opensynth.models.faraday.extended_model import ExtendedFaradayModel
 from opensynth.models.faraday.model import FaradayModel
 from opensynth.models.faraday.vae_model import Decoder, Encoder, FaradayVAE
 
-__all__ = ["FaradayVAE", "FaradayModel", "Encoder", "Decoder"]
+__all__ = [
+    "FaradayVAE",
+    "FaradayModel",
+    "ExtendedFaradayModel",
+    "Encoder",
+    "Decoder",
+]
